refactor: extracted data calculations to their dedicated file
